Refactor: Amazon and Mercadolibre - cards page removed - styles
