+++ src/pages/integrations/amazon/amazon.html
@@ -6,59 +6,72 @@
 
 </ion-header>
 
-<ion-content *ngIf="!giftCards">
+<ion-content *ngIf="!giftCards" no-bounce>
   <div class="box-notification warning" *ngIf="network == 'testnet'">
     Sandbox version. Only for testing purpose.
   </div>
-  <div class="integration-onboarding">
-    <div class="integration-onboarding-logo">
+  <div class="center-header">
+    <div class="logo">
       <img src="assets/img/amazon/GCs-logo-cllb.png" alt="Amazon.com Gift Card">
     </div>
-    <div class="integration-onboarding-description">Gift Cards are only redeemable on Amazon.com (US website). Cards never expire and can be redeemed towards millions of
+    <div class="description" padding>Gift Cards are only redeemable on Amazon.com (US website). Cards never expire and can be redeemed towards millions of
       items.
     </div>
-    <div class="integration-onboarding-cta" *ngIf="!showOauthForm">
-      <button ion-button (click)="goTo('Amount')" no-low-fee>Buy a Gift Card</button>
-      <button ion-button (click)="openExternalLink('https://www.amazon.com')">Visit Amazon.com &rarr;</button>
+    <div *ngIf="!showOauthForm">
+      <button ion-button outline block (click)="goTo('Amount')" no-low-fee>Buy a Gift Card</button>
+      <button ion-button clear small block color="dark" (click)="openExternalLink('https://www.amazon.com')">Visit Amazon.com &rarr;</button>
     </div>
   </div>
 </ion-content>
 
-<ion-content *ngIf="giftCards">
+<ion-content *ngIf="giftCards" no-bounce>
   <div class="box-notification warning" *ngIf="network == 'testnet'">
     Sandbox version. Only for testing purpose.
   </div>
 
   <div class="integration-giftCard-logo">
     <img src="assets/img/amazon/GCs-logo-cllb.png" alt="Amazon.com Gift Card">
-    <div>
-      <b>Only</b> redeemable on www.amazon.com (US website).</div>
+    <div class="subtitle">
+      <b>Only</b> redeemable on www.amazon.com (US website).
+    </div>
+    <button color="dark" ion-button clear icon-right no-low-fee (click)="goTo('Amount')">
+      Buy Gift Card
+      <ion-icon name="arrow-forward"></ion-icon>
+    </button>
   </div>
 
-  <ion-card>
-    <ion-card-content>
-      <ion-list>
-        <button ion-item (click)="goTo('Amount')" no-low-fee>
-          <ion-icon name="ios-pricetags-outline"></ion-icon>
-          <span>Buy Gift Card</span>
-        </button>
-        <button ion-item (click)="goTo('AmazonCards')">
-          <ion-icon name="ios-folder-outline"></ion-icon>
-          Your cards
-        </button>
-      </ion-list>
-    </ion-card-content>
-  </ion-card>
-
-  <div class="integration-giftCard-info">
-    *
-    <a (click)="openExternalLink('http://amazon.com')">Amazon.com</a> is not a sponsor of this promotion. Except as required by law,
-    <a (click)="openExternalLink('http://amazon.com')">Amazon.com</a> Gift Cards ("GCs") cannot be transferred for value or redeemed for cash. GCs may be used only for purchases
-    of eligible goods at
-    <a (click)="openExternalLink('http://amazon.com')">Amazon.com</a> or certain of its affiliated websites. For complete terms and conditions, see
-    <a (click)="openExternalLink('https://www.amazon.com/gc-legal')">www.amazon.com/gc-legal</a>. GCs are issued by ACI Gift Cards, Inc., a Washington corporation. All Amazon &reg;, &trade;
-    &amp; &copy; are IP of
-    <a (click)="openExternalLink('http://amazon.com')">Amazon.com</a>, Inc. or its affiliates. No expiration date or service fees.
-  </div>
-
+  <ion-list>
+    <ion-list-header color="light">Your cards</ion-list-header>
+    <button ion-item *ngFor="let item of giftCards | keys" (click)="openCardModal(item.value)">
+      <div class="amazon-card">
+        <img src="assets/img/amazon/a-smile_color_btn.png" alt="" class="avatar">
+        <ion-spinner *ngIf="updatingPending[item.value.invoiceId]" name="crescent"></ion-spinner>
+        <div class="card-info">
+          <h2 *ngIf="item.value.amount">
+            {{item.value.amount | currency : '$ ' : 2}} {{item.value.currency}}
+          </h2>
+          <p>
+            <span class="assertive" *ngIf="item.value.status == 'FAILURE' || item.value.status == 'RESEND'">Error</span>
+            <span class="assertive" *ngIf="item.value.status == 'expired'">Expired</span>
+            <span class="assertive" *ngIf="item.value.status == 'invalid'">Still waiting confirmation
+              <br> (Use higher fees setting to faster delivery)</span>
+            <span class="text-gray" *ngIf="item.value.status == 'PENDING'">Pending to confirmation</span>
+            <span class="assertive" *ngIf="item.value.status == 'SUCCESS' && item.value.cardStatus == 'Canceled'">Canceled</span>
+            <span class="text-gray">{{item.value.date | amTimeAgo}}</span>
+          </p>
+        </div>
+      </div>
+    </button>
+  </ion-list>
 </ion-content>
+
+<ion-footer *ngIf="giftCards" class="integration-giftCard-info">
+  *
+  <a (click)="openExternalLink('http://amazon.com')">Amazon.com</a> is not a sponsor of this promotion. Except as required by law,
+  <a (click)="openExternalLink('http://amazon.com')">Amazon.com</a> Gift Cards ("GCs") cannot be transferred for value or redeemed for cash. GCs may be used only for purchases
+  of eligible goods at
+  <a (click)="openExternalLink('http://amazon.com')">Amazon.com</a> or certain of its affiliated websites. For complete terms and conditions, see
+  <a (click)="openExternalLink('https://www.amazon.com/gc-legal')">www.amazon.com/gc-legal</a>. GCs are issued by ACI Gift Cards, Inc., a Washington corporation. All Amazon &reg;, &trade;
+  &amp; &copy; are IP of
+  <a (click)="openExternalLink('http://amazon.com')">Amazon.com</a>, Inc. or its affiliates. No expiration date or service fees.
+</ion-footer>
