+++ src/pages/integrations/amazon/buy-amazon/buy-amazon.html
@@ -26,11 +26,13 @@
       <button ion-item (click)="showWallets()" class="wallets-list">
         <div translate>From</div>
         <div class="wallet">
-          <img *ngIf="network == 'testnet'" [ngStyle]="{'background-color': wallet && wallet.color}" src="assets/img/icon-wallet-testnet.svg"
-            class="icon-wallet" />
-          <img *ngIf="network != 'testnet'" [ngStyle]="{'background-color': wallet && wallet.color}" src="assets/img/icon-wallet.svg"
-            class="icon-wallet" />
-          <span>{{wallet ? wallet.name : '...'}}</span>
+          <ion-icon item-start>
+            <img *ngIf="network == 'testnet'" [ngStyle]="{'background-color': wallet && wallet.color}" src="assets/img/icon-wallet-testnet.svg"
+              class="icon-wallet" />
+            <img *ngIf="network != 'testnet'" [ngStyle]="{'background-color': wallet && wallet.color}" src="assets/img/icon-wallet.svg"
+              class="icon-wallet" />
+          </ion-icon>
+          <div>{{wallet ? wallet.name : '...'}}</div>
         </div>
       </button>
       <div *ngIf="totalAmountStr">
