+++ src/pages/integrations/amazon/buy-amazon/buy-amazon.scss
@@ -25,17 +25,17 @@ page-buy-amazon {
       display: flex;
       align-items: center;
       margin-top: 1rem;
-      img {
-        width: 2.3rem;
-        margin-right: 1.5rem; 
-        border-radius: 0.3rem;
+      .icon-wallet {
         background-color: color($colors, primary);
+        margin-right: 1rem;
       }
     }
   }
   .amazon-info {
     * {
       white-space: normal;
+      font-size: 12px;
+      line-height: 16px;
     }
   }
 }
