+++ src/pages/integrations/mercado-libre/buy-mercado-libre/buy-mercado-libre.html
@@ -15,9 +15,6 @@
       </div>
       <div class="amount-label">
         <div class="amount">{{amountUnitStr}}</div>
-        <div class="alternative">
-          {{limitPerDayMessage}}
-        </div>
       </div>
     </ion-item>
 
@@ -25,10 +22,16 @@
       <button ion-item (click)="showWallets()" class="wallets-list">
         <div translate>From</div>
         <div class="wallet">
-          <img src="assets/img/icon-wallet.svg">
-          <span>{{wallet ? wallet.name : '...'}}</span>
+          <ion-icon item-start>
+            <img *ngIf="network == 'testnet'" [ngStyle]="{'background-color': wallet && wallet.color}" src="assets/img/icon-wallet-testnet.svg"
+              class="icon-wallet" />
+            <img *ngIf="network != 'testnet'" [ngStyle]="{'background-color': wallet && wallet.color}" src="assets/img/icon-wallet.svg"
+              class="icon-wallet" />
+          </ion-icon>
+          <div>{{wallet ? wallet.name : '...'}}</div>
         </div>
       </button>
+
       <div *ngIf="totalAmountStr">
         <ion-item-divider color="light">
           Details
