+++ src/pages/integrations/mercado-libre/mercado-libre-card-details/mercado-libre-card-details.scss
@@ -1,14 +1,24 @@
 page-mercado-libre-card-details {
-  .scroll-content {
-    padding: 2rem;
-  }
   .header-modal {
     text-align: center;
+    padding: 2rem;
     img {
       width: 230px;
     }
     &-amount {
       padding: 1rem;
+    }
+    .redeem {
+      margin-top: 2rem;
+    }
+    .invoice {
+      margin: 1.5rem 0 3rem 0;
+    }
+  }
+  .links {
+    text-align: center;
+    span {
+      color: color($colors, primary);
     }
   }
 }
