--- src/pages/integrations/mercado-libre/mercado-libre-cards/mercado-libre-cards.html
@@ -1,34 +1,0 @@
-<ion-header>
-
-  <ion-navbar>
-    <ion-title>{{'Your Gift Cards'|translate}}</ion-title>
-  </ion-navbar>
-
-</ion-header>
-
-
-<ion-content>
-  <ion-card>
-    <ion-list *ngIf="!hideCards">
-      <button ion-item *ngFor="let item of giftCards | keys" (click)="openCardModal(item.value)">
-          <div class="mercado-libre-card">
-            <img src="assets/img/mercado-libre/meli-card-24px.png" alt="" class="avatar">
-            <div class="card-info">
-              <h2 *ngIf="item.value.amount">
-                {{item.value.amount | currency : '$ ' : 2}} {{item.value.currency}}
-              </h2>
-              <span>
-                <span class="assertive" *ngIf="item.value.status == 'FAILURE'">Error</span>
-                <span class="dark" *ngIf="item.value.status == 'expired'">Invoice expired</span>
-                <span class="calm" *ngIf="item.value.status == 'invalid'">Still pending</span>
-                <span class="positive" *ngIf="item.value.status == 'PENDING'">Pending</span>
-                <span class="assertive" *ngIf="item.value.cardStatus == 'inactive'" translate>Inactive</span>
-                <span class="assertive" *ngIf="item.value.cardStatus == 'expired'" translate>Expired</span>
-              </span>
-              <span class="dark">{{item.value.date | amTimeAgo}}</span>
-            </div>
-          </div>
-        </button>
-    </ion-list>
-  </ion-card>
-</ion-content>
