+++ src/pages/integrations/mercado-libre/mercado-libre.html
@@ -6,47 +6,62 @@
 
 </ion-header>
 
-<ion-content *ngIf="!giftCards">
+<ion-content *ngIf="!giftCards" no-bounce>
   <div class="box-notification warning" *ngIf="network == 'testnet'">
     Sandbox version. Only for testing purpose.
   </div>
-  <div class="integration-onboarding">
-    <div class="integration-onboarding-logo">
+  <div class="center-header">
+    <div class="logo">
       <img src="assets/img/mercado-libre/mlbr.svg" alt="Mercado Libre">
     </div>
-    <div class="integration-onboarding-description" translate>
+    <div class="description" translate padding>
       <b>Only</b> redeemable on Mercado Livre (Brazil)
     </div>
-    <div class="integration-onboarding-cta" *ngIf="!showOauthForm">
-      <button ion-button (click)="goTo('Amount')" no-low-fee translate>Buy a Gift Card</button>
-      <button ion-button (click)="openExternalLink('https://www.mercadolivre.com.br')" translate>Visit mercadolivre.com.br &rarr;</button>
+    <div *ngIf="!showOauthForm">
+      <button ion-button outline block (click)="goTo('Amount')" no-low-fee translate>Buy a Gift Card</button>
+      <button ion-button clear small block color="dark" (click)="openExternalLink('https://www.mercadolivre.com.br')" translate>Visit mercadolivre.com.br &rarr;</button>
     </div>
   </div>
 </ion-content>
 
-<ion-content *ngIf="giftCards">
+<ion-content *ngIf="giftCards" no-bounce>
   <div class="box-notification warning" *ngIf="network == 'testnet'">
     Sandbox version. Only for testing purpose.
   </div>
 
   <div class="integration-giftCard-logo">
     <img src="assets/img/mercado-libre/mlbr.svg" alt="Mercado Libre">
-    <div translate>
-      <b>Only</b> redeemable on Mercado Livre (Brazil)</div>
+    <div class="subtitle" translate>
+      <b>Only</b> redeemable on Mercado Livre (Brazil)
+    </div>
+    <button color="dark" ion-button clear icon-right no-low-fee (click)="goTo('Amount')">
+      <span translate>Buy Gift Card</span>
+      <ion-icon name="arrow-forward"></ion-icon>
+    </button>
   </div>
 
-  <ion-card>
-    <ion-card-content>
-      <ion-list>
-        <button ion-item no-low-fee (click)="goTo('Amount')">
-          <ion-icon name="ios-pricetags-outline"></ion-icon>
-          <span translate>Buy Gift Card</span>
-        </button>
-        <button ion-item (click)="goTo('MercadoLibreCards')">
-          <ion-icon name="ios-folder-outline"></ion-icon>
-          <span translate>Your Gift Cards</span>
-        </button>
-      </ion-list>
-    </ion-card-content>
-  </ion-card>
+  <ion-list>
+    <ion-list-header color="light">{{'Your Gift Cards'|translate}}</ion-list-header>
+    <button ion-item *ngFor="let item of giftCards | keys" (click)="openCardModal(item.value)">
+      <div class="mercado-libre-card">
+        <img src="assets/img/mercado-libre/meli-card-24px.png" alt="" class="avatar">
+        <div class="card-info">
+          <div class="amount">
+            <h2 *ngIf="item.value.amount">
+              {{item.value.amount | currency : '$ ' : 2}} {{item.value.currency}}
+            </h2>
+            <span>
+              <span class="assertive" *ngIf="item.value.status == 'FAILURE'">Error</span>
+              <span class="dark" *ngIf="item.value.status == 'expired'">Invoice expired</span>
+              <span class="calm" *ngIf="item.value.status == 'invalid'">Still pending</span>
+              <span class="positive" *ngIf="item.value.status == 'PENDING'">Pending</span>
+              <span class="assertive" *ngIf="item.value.cardStatus == 'inactive'" translate>Inactive</span>
+              <span class="assertive" *ngIf="item.value.cardStatus == 'expired'" translate>Expired</span>
+            </span>
+          </div>
+          <span class="dark">{{item.value.date | amTimeAgo}}</span>
+        </div>
+      </div>
+    </button>
+  </ion-list>
 </ion-content>
